move formatters for individual headers/parameters to extension base traits
