+++ src/main/scala/nl/knaw/dans/lib/logging/servlet/masked/request/MaskedAuthorizationHeader.scala
@@ -15,10 +15,10 @@
  */
 package nl.knaw.dans.lib.logging.servlet.masked.request
 
-import nl.knaw.dans.lib.logging.servlet.{ HeaderMapEntry, RequestLogFormatter }
+import nl.knaw.dans.lib.logging.servlet.{ HeaderMapEntry, RequestLogExtensionBase }
 import org.scalatra.ScalatraBase
 
-private[masked] trait MaskedAuthorizationHeader extends RequestLogFormatter {
+private[masked] trait MaskedAuthorizationHeader extends RequestLogExtensionBase {
   this: ScalatraBase =>
 
   abstract override protected def formatHeader(header: HeaderMapEntry): HeaderMapEntry = {
